requirements.txt
django-ninja>=0.17.0
django>=3.2
pydantic>=1.0